+++ source/client/source/Framework/Rendering/ImGuiWindows/RenderingWindow.cs
@@ -57,7 +57,7 @@ public class RenderingWindow : ImGuiWindow
 
         ImGui.Separator();
         
-        ulong loadedChunks = RenderingStats.LoadedColumnCount * Constants.CHUNK_COLUMN_HEIGHT;
+        uint loadedChunks = (uint)RenderingStats.LoadedColumnCount * Constants.CHUNK_COLUMN_HEIGHT;
         ImGui.Text($"Loaded Blocks = {(loadedChunks * Constants.CHUNK_SIZE_CUBED).ToString("#,0", _numberFormat)}");
         ImGui.Text($"Loaded Chunks = {loadedChunks.ToString("#,0", _numberFormat)}");
         ImGui.Text($"Loaded Columns = {RenderingStats.LoadedColumnCount}");
